fix image get

+++ src/main/kotlin/com/example/demo/datasource/EventRepo.kt
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface EventRepo : JpaRepository<Event, Long> {
+interface EventRepo : JpaRepository<Event, Int> {
 }
